combine feature-store with feature for CLI
