+++ src/zenml/cli/__init__.py
@@ -815,7 +815,6 @@ zenml model-deployer logs MODEL_DEPLOYER_NAME
 from zenml.cli.base import *  # noqa
 from zenml.cli.config import *  # noqa
 from zenml.cli.example import *  # noqa
-from zenml.cli.feature import *  # noqa
 from zenml.cli.integration import *  # noqa
 from zenml.cli.pipeline import *  # noqa
 from zenml.cli.served_models import *  # noqa
