+++ src/zenml/cli/stack_components.py
@@ -13,6 +13,7 @@
 #  permissions and limitations under the License.
 """Functionality to generate stack component CLI commands."""
 
+import getpass
 import time
 from importlib import import_module
 from typing import (
@@ -33,32 +34,9 @@ from rich.markdown import Markdown
 
 from zenml.cli import utils as cli_utils
 from zenml.cli.cli import TagGroup, cli
-from zenml.config.global_config import GlobalConfiguration
-from zenml.console import console
-from zenml.constants import MANDATORY_COMPONENT_ATTRIBUTES
-from zenml.enums import CliCategories, StackComponentType
-from zenml.exceptions import EntityExistsError
-from zenml.io import fileio
-from zenml.repository import Repository
-from zenml.stack import StackComponent
-from zenml.utils.analytics_utils import AnalyticsEvent, track_event
-from zenml.utils.source_utils import validate_flavor_source
-from zenml.zen_stores.models.component_wrapper import ComponentWrapper
-
-if TYPE_CHECKING:
-    from zenml.annotators.base_annotator import BaseAnnotator
-    from zenml.secrets_managers.base_secrets_manager import BaseSecretsManager
-
-
-import getpass
-from typing import TYPE_CHECKING, List
-
-import click
-from pydantic import ValidationError
-
-from zenml.cli.cli import TagGroup, cli
 from zenml.cli.utils import (
     confirmation,
+    declare,
     error,
     expand_argument_value_from_file,
     parse_unknown_options,
@@ -66,11 +44,23 @@ from zenml.cli.utils import (
     print_list_items,
     warning,
 )
+from zenml.config.global_config import GlobalConfiguration
 from zenml.console import console
+from zenml.constants import MANDATORY_COMPONENT_ATTRIBUTES
 from zenml.enums import CliCategories, StackComponentType
-from zenml.exceptions import SecretExistsError
+from zenml.exceptions import EntityExistsError, SecretExistsError
+from zenml.io import fileio
 from zenml.repository import Repository
 from zenml.secret import ARBITRARY_SECRET_SCHEMA_TYPE
+from zenml.stack import StackComponent
+from zenml.utils.analytics_utils import AnalyticsEvent, track_event
+from zenml.utils.source_utils import validate_flavor_source
+from zenml.zen_stores.models.component_wrapper import ComponentWrapper
+
+if TYPE_CHECKING:
+    from zenml.annotators.base_annotator import BaseAnnotator
+    from zenml.feature_stores.base_feature_store import BaseFeatureStore
+    from zenml.secrets_managers.base_secrets_manager import BaseSecretsManager
 
 
 def _get_required_attributes(
@@ -1868,6 +1858,102 @@ def register_secrets_manager_subcommands() -> None:
                 console.print("Deleted all secrets.")
 
 
+def register_feature_store_subcommands() -> None:
+    """Registers CLI subcommands for the Feature Store."""
+    feature_store_group = cast(TagGroup, cli.commands.get("feature-store"))
+    if feature_store_group:
+
+        @feature_store_group.group(
+            cls=TagGroup,
+            help="Commands for interacting with your features.",
+        )
+        @click.pass_context
+        def feature(ctx: click.Context) -> None:
+            """Features as obtained from a feature store.
+
+            Args:
+                ctx: The click context.
+            """
+            repo = Repository()
+            active_stack = repo.zen_store.get_stack(name=repo.active_stack_name)
+            feature_store_wrapper = active_stack.get_component_wrapper(
+                StackComponentType.FEATURE_STORE
+            )
+            if feature_store_wrapper is None:
+                error(
+                    "No active feature store found. Please create a feature store "
+                    "first and add it to your stack."
+                )
+                return
+            ctx.obj = feature_store_wrapper.to_component()
+
+        @feature.command("get-data-sources")
+        @click.pass_obj
+        def get_data_sources(feature_store: "BaseFeatureStore") -> None:
+            """Get all data sources from the feature store.
+
+            Args:
+                feature_store: The feature store.
+            """
+            data_sources = feature_store.get_data_sources()
+            declare(f"Data sources: {data_sources}")
+
+        @feature.command("get-entities")
+        @click.pass_obj
+        def get_entities(feature_store: "BaseFeatureStore") -> None:
+            """Get all entities from the feature store.
+
+            Args:
+                feature_store: The feature store.
+            """
+            entities = feature_store.get_entities()
+            declare(f"Entities: {entities}")
+
+        @feature.command("get-feature-services")
+        @click.pass_obj
+        def get_feature_services(feature_store: "BaseFeatureStore") -> None:
+            """Get all feature services from the feature store.
+
+            Args:
+                feature_store: The feature store.
+            """
+            feature_services = feature_store.get_feature_services()
+            declare(f"Feature services: {feature_services}")
+
+        @feature.command("get-feature-views")
+        @click.pass_obj
+        def get_feature_views(feature_store: "BaseFeatureStore") -> None:
+            """Get all feature views from the feature store.
+
+            Args:
+                feature_store: The feature store.
+            """
+            feature_views = feature_store.get_feature_views()
+            declare(f"Feature views: {feature_views}")
+
+        @feature.command("get-project")
+        @click.pass_obj
+        def get_project(feature_store: "BaseFeatureStore") -> None:
+            """Get the current project name from the feature store.
+
+            Args:
+                feature_store: The feature store.
+            """
+            project = feature_store.get_project()
+            declare(f"Project name: {project}")
+
+        @feature.command("get-feast-version")
+        @click.pass_obj
+        def get_feast_version(feature_store: "BaseFeatureStore") -> None:
+            """Get the current Feast version being used.
+
+            Args:
+                feature_store: The feature store.
+            """
+            version = feature_store.get_feast_version()
+            declare(f"Feast version: {version}")
+
+
 def register_annotator_subcommands() -> None:
     """Registers CLI subcommands for the annotator."""
     annotator_group = cast(TagGroup, cli.commands.get("annotator"))
@@ -2019,6 +2105,8 @@ def register_all_stack_component_cli_commands() -> None:
             register_annotator_subcommands()
         elif component_type == StackComponentType.SECRETS_MANAGER:
             register_secrets_manager_subcommands()
+        elif component_type == StackComponentType.FEATURE_STORE:
+            register_feature_store_subcommands()
 
 
 register_all_stack_component_cli_commands()
